+++ test-data/projects/shadow-w-backend/src/main/server.clj
@@ -1,0 +1,7 @@
+(ns main.server
+  (:gen-class))
+
+(defn -main
+  "I don't do a whole lot ... yet."
+  [& _args]
+  (println "Hello, World!"))